requirements.txt
requests
pandas
pyarrow
dlt[duckdb]
python-dotenv